More UI elements, especially health bars.
This includes full and empty health bars, the health section, and the health section that fits at the end of the bar (when the player is almost out of life). I also added the "X" icon showing how many bullets the player has, and the pistol's bullet icon; both of these are designed to go on the display.